Makefile: Rework liberty timing generation logic
This now uses the python script to list all possible liberty files
that can be generated for all libraries that are checked out.

Each has its own target so parallel build work as expected and each
has its own dependency list that should trigger a rebuild if any of
the source file is updated.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>